requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: win-64
absl-py=2.1.0=pypi_0
aniso8601=9.0.1=pypi_0
anyio=4.2.0=py310haa95532_0
argon2-cffi=21.3.0=pyhd3eb1b0_0
argon2-cffi-bindings=21.2.0=py310h2bbff1b_0
asttokens=2.0.5=pyhd3eb1b0_0
astunparse=1.6.3=pypi_0
async-lru=2.0.4=py310haa95532_0
attrs=23.1.0=py310haa95532_0
babel=2.11.0=py310haa95532_0
beautifulsoup4=4.12.3=py310haa95532_0
blas=1.0=mkl
bleach=4.1.0=pyhd3eb1b0_0
blinker=1.8.2=pypi_0
bottleneck=1.3.7=py310h9128911_0
brotli=1.0.9=h2bbff1b_8
brotli-bin=1.0.9=h2bbff1b_8
brotli-python=1.0.9=py310hd77b12b_8
bzip2=1.0.8=h2bbff1b_6
ca-certificates=2024.7.2=haa95532_0
cachetools=5.5.0=pypi_0
certifi=2024.8.30=py310haa95532_0
cffi=1.16.0=py310h2bbff1b_1
charset-normalizer=3.3.2=pyhd3eb1b0_0
click=8.1.7=pypi_0
colorama=0.4.6=py310haa95532_0
comm=0.2.1=py310haa95532_0
contourpy=1.2.0=py310h59b6b97_0
cycler=0.11.0=pyhd3eb1b0_0
debugpy=1.6.7=py310hd77b12b_0
decorator=5.1.1=pyhd3eb1b0_0
defusedxml=0.7.1=pyhd3eb1b0_0
exceptiongroup=1.2.0=py310haa95532_0
executing=0.8.3=pyhd3eb1b0_0
flask=3.0.3=pypi_0
flask-restful=0.3.10=pypi_0
flatbuffers=24.3.25=pypi_0
fonttools=4.51.0=py310h2bbff1b_0
freetype=2.12.1=ha860e81_0
gast=0.6.0=pypi_0
google-auth=2.34.0=pypi_0
google-auth-oauthlib=1.2.1=pypi_0
google-pasta=0.2.0=pypi_0
grpcio=1.66.1=pypi_0
h5py=3.11.0=pypi_0
icc_rt=2022.1.0=h6049295_2
icu=73.1=h6c2663c_0
idna=3.7=py310haa95532_0
intel-openmp=2023.1.0=h59b6b97_46320
ipykernel=6.28.0=py310haa95532_0
ipython=8.25.0=py310haa95532_0
itsdangerous=2.2.0=pypi_0
jedi=0.19.1=py310haa95532_0
jinja2=3.1.4=py310haa95532_0
joblib=1.4.2=py310haa95532_0
jpeg=9e=h827c3e9_3
json5=0.9.6=pyhd3eb1b0_0
jsonschema=4.19.2=py310haa95532_0
jsonschema-specifications=2023.7.1=py310haa95532_0
jupyter-lsp=2.2.0=py310haa95532_0
jupyter_client=8.6.0=py310haa95532_0
jupyter_core=5.7.2=py310haa95532_0
jupyter_events=0.10.0=py310haa95532_0
jupyter_server=2.14.1=py310haa95532_0
jupyter_server_terminals=0.4.4=py310haa95532_1
jupyterlab=4.0.11=py310haa95532_0
jupyterlab_pygments=0.1.2=py_0
jupyterlab_server=2.25.1=py310haa95532_0
keras=2.15.0=pypi_0
kiwisolver=1.4.4=py310hd77b12b_0
krb5=1.20.1=h5b6d351_0
lcms2=2.12=h83e58a3_0
lerc=3.0=hd77b12b_0
libbrotlicommon=1.0.9=h2bbff1b_8
libbrotlidec=1.0.9=h2bbff1b_8
libbrotlienc=1.0.9=h2bbff1b_8
libclang=18.1.1=pypi_0
libclang13=14.0.6=default_h8e68704_1
libdeflate=1.17=h2bbff1b_1
libffi=3.4.4=hd77b12b_1
libpng=1.6.39=h8cc25b3_0
libpq=12.17=h906ac69_0
libsodium=1.0.18=h62dcd97_0
libtiff=4.5.1=hd77b12b_0
libwebp-base=1.3.2=h2bbff1b_0
lz4-c=1.9.4=h2bbff1b_1
markdown=3.7=pypi_0
markdown-it-py=3.0.0=pypi_0
markupsafe=2.1.3=py310h2bbff1b_0
matplotlib=3.9.2=py310haa95532_0
matplotlib-base=3.9.2=py310he19b0ae_0
matplotlib-inline=0.1.6=py310haa95532_0
mdurl=0.1.2=pypi_0
mistune=2.0.4=py310haa95532_0
mkl=2023.1.0=h6b88ed4_46358
mkl-service=2.4.0=py310h2bbff1b_1
mkl_fft=1.3.8=py310h2bbff1b_0
mkl_random=1.2.4=py310h59b6b97_0
ml-dtypes=0.2.0=pypi_0
namex=0.0.8=pypi_0
nbclient=0.8.0=py310haa95532_0
nbconvert=7.10.0=py310haa95532_0
nbformat=5.9.2=py310haa95532_0
nest-asyncio=1.6.0=py310haa95532_0
notebook=7.0.8=py310haa95532_2
notebook-shim=0.2.3=py310haa95532_0
numexpr=2.8.7=py310h2cd9be0_0
numpy=1.26.4=py310h055cbcc_0
numpy-base=1.26.4=py310h65a83cf_0
oauthlib=3.2.2=pypi_0
opencv-python=4.10.0.84=pypi_0
openjpeg=2.5.2=hae555c5_0
openssl=3.0.15=h827c3e9_0
opt-einsum=3.3.0=pypi_0
optree=0.12.1=pypi_0
overrides=7.4.0=py310haa95532_0
packaging=24.1=py310haa95532_0
pandas=2.2.2=py310h5da7b33_0
pandocfilters=1.5.0=pyhd3eb1b0_0
parso=0.8.3=pyhd3eb1b0_0
pillow=10.4.0=py310h827c3e9_0
pip=24.2=py310haa95532_0
platformdirs=3.10.0=py310haa95532_0
ply=3.11=py310haa95532_0
prometheus_client=0.14.1=py310haa95532_0
prompt-toolkit=3.0.43=py310haa95532_0
prompt_toolkit=3.0.43=hd3eb1b0_0
protobuf=4.25.4=pypi_0
psutil=5.9.0=py310h2bbff1b_0
pure_eval=0.2.2=pyhd3eb1b0_0
pyasn1=0.6.0=pypi_0
pyasn1-modules=0.4.0=pypi_0
pybind11-abi=5=hd3eb1b0_0
pycparser=2.21=pyhd3eb1b0_0
pygments=2.15.1=py310haa95532_1
pyparsing=3.1.2=py310haa95532_0
pyqt=5.15.10=py310hd77b12b_0
pyqt5-sip=12.13.0=py310h2bbff1b_0
pysocks=1.7.1=py310haa95532_0
python=3.10.14=he1021f5_1
python-dateutil=2.9.0post0=py310haa95532_2
python-fastjsonschema=2.16.2=py310haa95532_0
python-json-logger=2.0.7=py310haa95532_0
python-tzdata=2023.3=pyhd3eb1b0_0
pytz=2024.1=py310haa95532_0
pywin32=305=py310h2bbff1b_0
pywinpty=2.0.10=py310h5da7b33_0
pyyaml=6.0.1=py310h2bbff1b_0
pyzmq=25.1.2=py310hd77b12b_0
qt-main=5.15.2=h19c9488_10
referencing=0.30.2=py310haa95532_0
requests=2.32.3=py310haa95532_0
requests-oauthlib=2.0.0=pypi_0
rfc3339-validator=0.1.4=py310haa95532_0
rfc3986-validator=0.1.1=py310haa95532_0
rich=13.8.0=pypi_0
rpds-py=0.10.6=py310h062c2fa_0
rsa=4.9=pypi_0
scikit-learn=1.5.1=py310hc64d2fc_0
scipy=1.13.1=py310h8640f81_0
seaborn=0.13.2=py310haa95532_0
send2trash=1.8.2=py310haa95532_0
setuptools=72.1.0=py310haa95532_0
sip=6.7.12=py310hd77b12b_0
six=1.16.0=pyhd3eb1b0_1
sniffio=1.3.0=py310haa95532_0
soupsieve=2.5=py310haa95532_0
sqlite=3.45.3=h2bbff1b_0
stack_data=0.2.0=pyhd3eb1b0_0
tbb=2021.8.0=h59b6b97_0
tensorboard=2.15.2=pypi_0
tensorboard-data-server=0.7.2=pypi_0
tensorflow=2.15.0=pypi_0
tensorflow-estimator=2.15.0=pypi_0
tensorflow-intel=2.15.0=pypi_0
tensorflow-io-gcs-filesystem=0.31.0=pypi_0
termcolor=2.4.0=pypi_0
terminado=0.17.1=py310haa95532_0
threadpoolctl=3.5.0=py310h9909e9c_0
tinycss2=1.2.1=py310haa95532_0
tk=8.6.14=h0416ee5_0
tomli=2.0.1=py310haa95532_0
tornado=6.4.1=py310h827c3e9_0
traitlets=5.14.3=py310haa95532_0
typing-extensions=4.11.0=py310haa95532_0
typing_extensions=4.11.0=py310haa95532_0
tzdata=2024a=h04d1e81_0
unicodedata2=15.1.0=py310h2bbff1b_0
urllib3=2.2.2=py310haa95532_0
vc=14.40=h2eaa2aa_0
vs2015_runtime=14.40.33807=h98bb1dd_0
wcwidth=0.2.5=pyhd3eb1b0_0
webencodings=0.5.1=py310haa95532_1
websocket-client=1.8.0=py310haa95532_0
werkzeug=3.0.4=pypi_0
wheel=0.43.0=py310haa95532_0
win_inet_pton=1.1.0=py310haa95532_0
winpty=0.4.3=4
wrapt=1.14.1=pypi_0
xz=5.4.6=h8cc25b3_1
yaml=0.2.5=he774522_0
zeromq=4.3.5=hd77b12b_0
zlib=1.2.13=h8cc25b3_1
zstd=1.5.5=hd43e919_2
